+++ packages/jest-config/src/defaults.js
@@ -36,6 +36,7 @@ export default ({
   clearMocks: false,
   coveragePathIgnorePatterns: [NODE_MODULES_REGEXP],
   coverageReporters: ['json', 'text', 'lcov', 'clover'],
+  detectLeaks: false,
   expand: false,
   globals: {},
   haste: {
@@ -55,6 +56,7 @@ export default ({
   runner: 'jest-runner',
   snapshotSerializers: [],
   testEnvironment: 'jest-environment-jsdom',
+  testEnvironmentOptions: {},
   testFailureExitCode: 1,
   testLocationInResults: false,
   testMatch: ['**/__tests__/**/*.js?(x)', '**/?(*.)(spec|test).js?(x)'],